lumina-portfolio/backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlmodel==0.0.21
pillow==12.0.0
pillow-heif==0.17.0
python-multipart==0.0.9
pydantic-settings==2.6.1

